feat: enhance test suite for ATT&CK to Excel export functionality
- Added slow marker to several tests in `test_cli.py` and `test_layers.py` to categorize them for longer execution times.
- Refactored `test_export_excel` in `test_cli.py` to utilize a mock `ToExcel` class for testing Excel export argument wiring.
- Improved `test_export_with_memstore_uses_current_dataframe_builder` and `test_export_with_pre_v18_version_uses_legacy_dataframe_builder` in `test_to_excel.py` to validate dataframe building logic based on version.
- Introduced new tests in `test_to_excel.py` to verify the creation of expected workbooks, handling of empty data, deduplication of citations, and sanitization of matrix sheet names.
- Updated `uv.lock` to require pytest version 9.0 or higher for development dependencies.
