+++ docs/CONTRIBUTING.md
@@ -51,10 +51,13 @@ Run `just` with no arguments to see all available commands. Here are the most co
 
 ```bash
 just lint          # Run pre-commit hooks (ruff format) on all files
-just test          # Run tests
+just test          # Run the full test suite, matching CI expectations
+just test-fast     # Run the fast local subset, excluding integration and slow tests
 just test-cov      # Run tests with coverage report
 just build         # Build the package
 ```
+
+Use `just test-fast` while iterating locally on changes that do not need full STIX-backed export or other slow integration coverage. Tests or setup steps that normally take longer than 10 seconds should be marked `slow`, so they are skipped by `just test-fast`. Before opening a PR, run `just test`; GitHub Actions also runs the full suite with coverage.
 
 To run STIX-backed tests against specific local bundles, pass the bundle paths to pytest:
