+++ tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Shared pytest fixture plugins for mitreattack-python tests."""
